Factored InvitationService out of ConnectionService
