+++ packages/sdk/src/services/connection-service.ts
@@ -1,9 +1,10 @@
-import { Connection, ConnectionApi, ConnectionsResponse, Invitation } from '@meeco/vault-api-sdk';
+import { Connection, ConnectionApi, ConnectionsResponse } from '@meeco/vault-api-sdk';
 import { AuthData } from '../models/auth-data';
 import { ConnectionCreateData } from '../models/connection-create-data';
 import { EncryptionKey } from '../models/encryption-key';
 import { findConnectionBetween } from '../util/find-connection-between';
 import { getAllPaged, resultHasNext } from '../util/paged';
+import { InvitationService } from './invitation-service';
 import Service from './service';
 
 export interface IDecryptedConnection {
@@ -19,60 +20,9 @@ export class ConnectionService extends Service<ConnectionApi> {
     return this.vaultAPIFactory(vaultToken).ConnectionApi;
   }
 
-  public async createInvitation(name: string, auth: AuthData): Promise<Invitation> {
-    const keyPair = await this.createAndStoreKeyPair(
-      auth.keystore_access_token,
-      auth.key_encryption_key
-    );
-
-    this.logger.log('Encrypting recipient name');
-    const encryptedName: string = await this.encryptName(name, auth.data_encryption_key);
-
-    this.logger.log('Sending invitation request');
-    return await this.vaultAPIFactory(auth)
-      .InvitationApi.invitationsPost({
-        public_key: {
-          keypair_external_id: keyPair.keystoreStoredKeyPair.id,
-          public_key: keyPair.keyPair.publicKey,
-        },
-        invitation: {
-          encrypted_recipient_name: encryptedName,
-        },
-      })
-      .then(result => result.invitation);
-  }
-
-  public async acceptInvitation(
-    name: string,
-    invitationToken: string,
-    auth: AuthData
-  ): Promise<Connection> {
-    const keyPair = await this.createAndStoreKeyPair(
-      auth.keystore_access_token,
-      auth.key_encryption_key
-    );
-
-    this.logger.log('Encrypting connection name');
-    const encryptedName: string = await this.encryptName(name, auth.data_encryption_key);
-
-    this.logger.log('Accepting invitation');
-    return await this.getAPI(auth.vault_access_token)
-      .connectionsPost({
-        public_key: {
-          keypair_external_id: keyPair.keystoreStoredKeyPair.id,
-          public_key: keyPair.keyPair.publicKey,
-        },
-        connection: {
-          encrypted_recipient_name: encryptedName,
-          invitation_token: invitationToken,
-        },
-      })
-      .then(res => res.connection);
-  }
-
   /**
    * Note this only works if we have authentication data for both connecting users.
-   * For more typical use cases you should manually call {@link createInvitation}
+   * For more typical use cases you should manually call {@link InvitationService.create}
    * as one user and {@link acceptInvitation} as the other user.
    */
   public async createConnection(config: ConnectionCreateData) {
@@ -92,8 +42,10 @@ export class ConnectionService extends Service<ConnectionApi> {
       throw new Error('Connection exists between the specified users');
     }
 
-    const invitation = await this.createInvitation(options.toName, from);
-    await this.acceptInvitation(options.fromName, invitation.token, to);
+    const invitations = new InvitationService(this.environment);
+
+    const invitation = await invitations.create(options.toName, from);
+    await invitations.accept(options.fromName, invitation.token, to);
 
     // Now the connection has been created we need to re-fetch the original user's connection.
     // We might as well fetch both to ensure it's connected both ways correctly.
@@ -114,7 +66,7 @@ export class ConnectionService extends Service<ConnectionApi> {
   }
 
   /**
-   * @deprecated Use [list] instead.
+   * @deprecated Use {@link list} instead.
    * @param user
    */
   public async listConnections(user: AuthData) {
@@ -183,41 +135,5 @@ export class ConnectionService extends Service<ConnectionApi> {
       );
       return Promise.all(decryptions);
     });
-  }
-
-  private encryptName(name: string, dek: EncryptionKey) {
-    return Service.cryppo
-      .encryptWithKey({
-        data: name,
-        key: dek.key,
-        strategy: Service.cryppo.CipherStrategy.AES_GCM,
-      })
-      .then(result => result.serialized);
-  }
-
-  private async createAndStoreKeyPair(keystoreToken: string, keyEncryptionKey: EncryptionKey) {
-    this.logger.log('Generating key pair');
-    const keyPair = await Service.cryppo.generateRSAKeyPair();
-
-    const toPrivateKeyEncrypted = await Service.cryppo.encryptWithKey({
-      data: keyPair.privateKey,
-      key: keyEncryptionKey.key,
-      strategy: Service.cryppo.CipherStrategy.AES_GCM,
-    });
-
-    const keystoreStoredKeyPair = await this.keystoreAPIFactory(keystoreToken)
-      .KeypairApi.keypairsPost({
-        public_key: keyPair.publicKey,
-        encrypted_serialized_key: toPrivateKeyEncrypted.serialized,
-        // API will 500 without
-        metadata: {},
-        external_identifiers: [],
-      })
-      .then(result => result.keypair);
-
-    return {
-      keyPair,
-      keystoreStoredKeyPair,
-    };
   }
 }
